Add guest invite button to new coordinator dashboard
