+++ app/src/components/common/index.ts
@@ -3,3 +3,4 @@ export {ApplicationTrackerHeader} from './ApplicationTrackerHeader';
 export {Header} from './Header';
 export {AppLayout} from './AppLayout';
 export {AuthenticatedHeader} from './AuthenticatedHeader';
+export {GuestInviteButton} from './GuestInviteButton';
